refactor: "hooks" directory renamed to "filters"
